backend/requirements.txt
fastapi
uvicorn
pymongo
motor
PyPDF2
python-dotenv
python-multipart
